Installs dependencies if not yet installed or force installed. Otherwise just rruns

+++ start.sh
@@ -2,47 +2,40 @@
 
 #!/bin/bash
 
-# Install in LAPI_server folder
-INSTALL_PATH="LAPI_server"
-
-# If given a directory with the -p flag, install in that directory
-while getopts ":p:" opt; do
+# If given a -f argument to force (re)install
+FORCE_REINSTALL="false"
+while getopts ":f:" opt; do
   case $opt in
-    p)
-      echo "Installing in folder: $OPTARG"
-      INSTALL_PATH=$OPTARG
-      ;;
     :)
-      echo "Please specify a path when using the path (-p) flag. Exiting..." >&2
-      exit
-      ;;
-    \?)
-      echo "Invalid option: -$OPTARG" 
+      echo "Force installing..."
+      FORCE_REINSTALL="true"
       ;;
   esac
 done
 
-mkdir -p $INSTALL_PATH
-cd $INSTALL_PATH
+# If no .LAS_status file, or force reinstall is selected
+if [ "$FORCE_REINSTALL" == "true" ] || [ ! -f .LAS_status ]; then
+  # Install NodeJS (use https://gist.github.com/isaacs/579814)
 
-# Install NodeJS (use https://gist.github.com/isaacs/579814)
+  # Install npm (probably included with Node)
 
-# Install npm (probably included with Node)
+  # Install npm dependencies
+  echo Installing dependencies from npm
+  npm install
+  npm install webpack -g
 
-# Install npm dependencies
-echo Installing dependencies from npm
-npm install
-npm install webpack -g
+  # Add/update hidden status file
+  touch .LAS_status
+  echo "Latest Update:" > .LAS_status
+  echo `date` > .LAS_status
 
-# Add/update hidden status file
-touch .LAS_status
-echo "Latest Update:" > .LAS_status
-echo `date` > .LAS_status
-
-# Build bundles
-echo Bundling React components
-webpack
+  # Build bundles
+  echo Bundling React components
+  webpack
+fi
 
 # Start server
 echo Starting LiveAPI Server
 npm start
+
+exit
